Add sbt v0.13 support and `sourcegraphEnable` command
Previously, sbt-sourcegraph only worked with sbt v1.3+ and it required
users to manually enable the SemanticDB compiler plugin. Now, the plugin
works with sbt v0.13.17 and v1.2.1+ and provides the `sourcegraphEnable`
command to automatically configure SemanticDB for both Java and Scala.

This change makes it much easier for users to integrate with
sbt-sourcegraph. For most sbt builds, the command
`sbt sourcegraphEnable sourcegraphLsif` will now reliably index
the repository.

The `sourcegraphEnable` command is an adaptation of the battle-tested
`scalafixEnable` command from sbt-scalafix with a few modifications:
- it automatically uses older versions of semanticdb-scalac instead of
  upgrading the Scala version.
- it configures the semanticdb-javac compiler plugin for hybrid
  Scala/Java projects.